requirements.txt
pandas==2.0.0
pyodbc==4.0.39
python-dotenv==1.0.0
SQLAlchemy==2.0.9
psycopg2==2.9.6